pyproject.toml
[project]
name = "matsuzawa-20260313"
version = "0.1.0"
description = "Adaptive Budgeted Chain-of-Thought experiment on GSM8K"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "datasets>=2.16.0",
    "google-generativeai>=0.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
